[Space-saver]: Re-structure
Add result to test-list generated

+++ dut_plugins/chromite_verilator_plugin/chromite_verilator_plugin.py
@@ -179,15 +179,19 @@ class chromite_verilator_plugin(object):
         return report_file_name
 
     @dut_hookimpl
-    def post_run(self, test_dict):
-        logger.debug("Going to remove stuff now")
-        for test in test_dict:
-            if test_dict[test]['result']:
-                logger.info("Removing extra files")
-                work_dir = test_dict[test]['work_dir']
-                os.remove(work_dir + '/app.log')
-                os.remove(work_dir + '/code.mem')
-                os.remove(work_dir + '/dut.disass')
-                os.remove(work_dir + '/dut.dump')
-                os.remove(work_dir + '/signature')
-        return
+    def post_run(self, test_dict, config):
+        if str_2_bool(config['river_core']['space_saver']):
+            logger.debug("Going to remove stuff now")
+            for test in test_dict:
+                if test_dict[test]['result'] and not test_dict[test]['result'] == 'Unavailable':
+                    logger.info("Removing extra files")
+                    work_dir = test_dict[test]['work_dir']
+                    try:
+                        os.remove(work_dir + '/app.log')
+                        os.remove(work_dir + '/code.mem')
+                        os.remove(work_dir + '/dut.disass')
+                        os.remove(work_dir + '/dut.dump')
+                        os.remove(work_dir + '/signature')
+                    except:
+                        logger.info(
+                            "Something went wrong trying to remove the files")
