+++ reference_plugins/spike_plugin/spike_plugin.py
@@ -128,11 +128,17 @@ class spike_plugin(object):
         return report_file_name
 
     @dut_hookimpl
-    def post_run(self, test_dict):
-        logger.debug("Going to remove stuff now")
-        for test in test_dict:
-            if test_dict[test]['result']:
-                logger.info("Removing extra files")
-                work_dir = test_dict[test]['work_dir']
-                os.remove(work_dir + '/ref.disass')
-                os.remove(work_dir + '/ref.dump')
+    def post_run(self, test_dict, config):
+        if str_2_bool(config['river_core']['space_saver']):
+            logger.debug("Going to remove stuff now")
+            for test in test_dict:
+                if test_dict[test]['result'] and not test_dict[test][
+                        'result'] == 'Unavailable':
+                    logger.info("Removing extra files")
+                    work_dir = test_dict[test]['work_dir']
+                    try:
+                        os.remove(work_dir + '/ref.disass')
+                        os.remove(work_dir + '/ref.dump')
+                    except:
+                        logger.info(
+                            "Something went wrong trying to remove the files")
